+++ rules/Php72/NodeFactory/AnonymousFunctionFactory.php
@@ -301,17 +301,7 @@ final class AnonymousFunctionFactory
             return null;
         }
 
-        // reset original node, to allow the printer to re-use the expr
-        $paramDefaultExpr->setAttribute(AttributeKey::ORIGINAL_NODE, null);
-        $this->simpleCallableNodeTraverser->traverseNodesWithCallable(
-            $paramDefaultExpr,
-            static function (Node $node): Node {
-                $node->setAttribute(AttributeKey::ORIGINAL_NODE, null);
-                return $node;
-            }
-        );
-
-        return $paramDefaultExpr;
+        return $this->nodeFactory->createReprintedExpr($paramDefaultExpr);
     }
 
     /**
