+++ rules/Php80/Rector/ClassMethod/AddParamBasedOnParentClassMethodRector.php
@@ -214,7 +214,7 @@ CODE_SAMPLE
             $paramDefault = $parentClassMethodParam->default;
 
             if ($paramDefault instanceof Expr) {
-                $paramDefault = $this->resolveParamDefault($paramDefault);
+                $paramDefault = $this->nodeFactory->createReprintedExpr($paramDefault);
             }
 
             $paramName = $this->nodeNameResolver->getName($parentClassMethodParam);
@@ -237,21 +237,6 @@ CODE_SAMPLE
         }
 
         return $node;
-    }
-
-    private function resolveParamDefault(Expr $expr): Expr
-    {
-        // reset original node, to allow the printer to re-use the expr
-        $expr->setAttribute(AttributeKey::ORIGINAL_NODE, null);
-        $this->traverseNodesWithCallable(
-            $expr,
-            static function (Node $node): Node {
-                $node->setAttribute(AttributeKey::ORIGINAL_NODE, null);
-                return $node;
-            }
-        );
-
-        return $expr;
     }
 
     private function resolveParamType(Param $param): null|Identifier|Name|ComplexType
